+++ NEWS.md
@@ -3,6 +3,7 @@
 ### NEW FEATURES
 ### OTHER CHANGES
 ### BUG FIXES 
+* Made it possible to use `trajectory.analysis` with univariate response data, omitting vector correlations output.
 
 # CHANGES IN RRPP 0.6.0 (Minor Release)